pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-tracker-analyzer"
version = "1.0.0"
description = "Automated system for tracking and analyzing academic papers using LLM"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Paper Tracker Team"}
]
keywords = ["arxiv", "papers", "llm", "analysis", "research"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "arxiv>=2.1.0",
    "pdfplumber>=0.11.0",
    "requests>=2.31.0",
    "click>=8.1.7",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.92.1",
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "requests-mock>=1.11.0",
    "pylint>=3.0.3",
    "mypy>=1.7.1",
    "black>=23.12.1",
]

[project.scripts]
paper-tracker = "src.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
markers = [
    "property_test: marks tests as property-based tests (deselect with '-m \"not property_test\"')",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
]
